--- a/radis/test/io/test_hdf5.py
+++ b/radis/test/io/test_hdf5.py
@@ -7,9 +7,62 @@ Created on Tue Jan 26 20:36:38 2021
 
 import pytest
 
-from radis.io.hdf5 import hdf2df
+from radis.io.hdf5 import HDF5Manager, hdf2df
 from radis.io.hitemp import fetch_hitemp
 from radis.misc.config import getDatabankEntries
+
+
+@pytest.mark.fast
+def test_hdf5_io_engines(*args, **kwargs):
+    """Test different engines implemented in :py:class:`radis.io.hdf5.HDF5Manager`"""
+
+    import os
+    from os.path import exists
+
+    for file in ["test_pytables.h5", "test_h5py.h5"]:
+        if exists(file):
+            os.remove(file)
+
+    # Test data
+    import h5py
+    import numpy as np
+    import pandas as pd
+
+    df0 = pd.DataFrame({"a": np.arange(10) ** 2, "b": np.arange(10) ** 3})
+    metadata0 = {"some_metadata": True}
+    # ... a Pandas HDFStore file
+    df0.to_hdf("test_pytables.h5", "df")
+    # ... a h5py file with same content
+    with h5py.File("test_h5py.h5", "w") as f:
+        # group = f.create_group("df")
+        for c in df0.columns:
+            f.create_dataset(c, data=df0[c])
+
+    # Test Pytables engine : add_metadata ; load ; read_metadata
+    manager = HDF5Manager(engine="pytables")
+    manager.add_metadata("test_pytables.h5", metadata0)
+
+    df = manager.load("test_pytables.h5")
+    assert (df == df0).all().all()
+    assert manager.read_metadata("test_pytables.h5") == metadata0
+
+    # Test h5py engine : add_metadata ; load ; read_metadata
+    manager = HDF5Manager(engine="h5py")
+    manager.add_metadata("test_h5py.h5", metadata0)
+    df = manager.load("test_h5py.h5", key=None)
+    assert (df == df0).all().all()
+    assert manager.read_metadata("test_h5py.h5") == metadata0
+
+    # Test vaex engine : add_metadata ; load ; read_
+    # .. also able to read h5py files
+    manager = HDF5Manager(engine="vaex")
+    manager.add_metadata("test_h5py.h5", metadata0)
+    df = manager.load("test_h5py.h5", key=None)
+    # this time; df is a vaex DataFrame
+    # ... it keeps the file open so we should close the file handle
+    assert (df.to_pandas_df() == df0).all().all()
+    df.close()
+    assert manager.read_metadata("test_h5py.h5") == metadata0
 
 
 @pytest.mark.needs_connection
@@ -59,4 +112,5 @@ def test_local_hdf5_lines_loading(*args, **kwargs):
 
 
 if __name__ == "__main__":
+    test_hdf5_io_engines()
     test_local_hdf5_lines_loading()