<!-- Please be sure to check out our contributing guidelines,
https://github.com/radis/radis/blob/develop/CONTRIBUTING.md . -->

### Description
<!-- Provide a general description of what your pull request does. -->

Adding vaex to load HDF5 files. Proof-of-concept : [test_vaex.ipynb](https://github.com/radis/radis-benchmark/blob/master/manual_benchmarks/test_vaex.ipynb)

Built on top of #281 : 

- [x] implemented different engines in HDF5 cache files & generation : vaex and h5py
- [x] added vaex filtering when loading database with wmin, wmax, or isotpes
- [ ] vaex is the default library to load HITEMP .hdf5 files  👉  "pytables" remains the default for the moment
- [x] auto update of your existing .hdF5 databases

```python
from radis.io.hitemp import fetch_hitemp
fetch_hitemp("OH", engine="pytables")       #  (if testing : make sure there is no vaex-formatted hdf5 already)
import radis; radis.AUTO_UPDATE_DATABASE = True
fetch_hitemp("OH", engine="vaex")   #  auto-converts to vaex-compatible HDF5

```

- [x] re-add comparison with already registered database (if exists)



Will be merged / rebased on #281 . 

Fixes #239 #295 


### TODO : 

- [ ] add multiple isotope support when querying vaex files
- [ ] add an auto-update test 
- [ ] write the parsed HITEMP files in Vaex-compatible / H5py format (see e3a1d734) 
- [x] Feature idea : for CO2 / H2O (multiple files) : only parse & download the files needed. Use the metadata keys wavenum_min; wavenum_max (?) in the database to see if new files should be downloaded ?  Could be extended for CDSD-4000 later.

Identified problems : 
- process sometimes stuck in a Spyder Ipython console (while converting to VAEX). File is converted properly, though, but only appears after the console is closed.   Similar to #338 and https://github.com/spyder-ide/spyder/issues/16183
- Some HDF5 files not closed properly if a crash happens : https://github.com/pandas-dev/pandas/issues/8692




### Performance improvement : 

Old (full loading and partial loading of HITEMP CO)

```python

In [2]: %timeit fetch_hitemp('CO', engine='pytables')
4.84 s ± 77 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [3]: %timeit fetch_hitemp('CO', engine='pytables', load_wavenum_min=2000, load_wavenum_max=3000)
814 ms ± 28 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

New:

```python
In [2]: %timeit fetch_hitemp('CO', engine='vaex')
1.42 s ± 35.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [3]: %timeit fetch_hitemp('CO', engine='vaex', load_wavenum_min=2000, load_wavenum_max=3000)
519 ms ± 27.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

So 1.5 to 3.5x faster.