@erwanp 2021-08-11
Update.   Branch rebased on #281 


I still get good improvement for loading of CO : 

Old : 

```python
%timeit fetch_hitemp('CO', engine='pytables')
4.25 s ± 138 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit fetch_hitemp('CO', engine='pytables', load_wavenum_min=2000, load_wavenum_max=3000)
480 ms ± 32 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

New:  

```python
%timeit fetch_hitemp('CO', engine='vaex')
647 ms ± 28.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit fetch_hitemp('CO', engine='vaex', load_wavenum_min=2000, load_wavenum_max=3000)
164 ms ± 4.83 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

So 3-5x faster


Most importantly, vaex can read multiple files easily while Pandas pd.concat is quite slow, and has to open all files independantly. This is very well suited for the H2O & CO2 linelists (10-20 files) of #281     (and CDSD-4000 @CorentinGrimaldi )  

This is already implemented : example with HITEMP CO2 below :

---

```python
from radis import SpectrumFactory
from time import time

sf = SpectrumFactory(
    wavenum_min=2000,
    wavenum_max=6000,
    molecule="CO2",
    isotope=1,
    verbose=3)

print("\n\n🏃‍ 1) With PyTables :")
t0 = time()
sf.fetch_databank("hitemp", hdf5_engine='pytables')
print(f"\n\n🏃‍ Database initialized in {time()-t0:.1f}s")
s1 = sf.eq_spectrum(Tgas=3000, name=f'pytables ({time()-t0:.1f}s)')
print(f"\n\n🏃‍ Spectrum calculated in {time()-t0:.1f}s (all included)")

#%%
print("\n\n🏃‍ 2) With Vaex:")
t0 = time()
sf.fetch_databank("hitemp", hdf5_engine='vaex')
print(f"\n\n🏃‍ Database initialized in {time()-t0:.1f}s")
s2 = sf.eq_spectrum(Tgas=3000, name=f'vaex ({time()-t0:.1f}s)')
print(f"\n\n🏃‍ Spectrum calculated in {time()-t0:.1f}s (all included)")

#%% Compare
# ... basically just check all lines are there
from radis import plot_diff
plot_diff(s1, s2)

```

Results : 

```

🏃‍ 1) With PyTables :
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_01500-02000_HITEMP2010.h5 in 0.10s (5607 rows, 30 columns, 3.21 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_02000-02125_HITEMP2010.h5 in 2.37s (271930 rows, 30 columns, 155.54 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_02125-02250_HITEMP2010.h5 in 3.98s (463711 rows, 30 columns, 265.24 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_02250-02500_HITEMP2010.h5 in 1.90s (228027 rows, 30 columns, 130.43 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_02500-03000_HITEMP2010.h5 in 0.64s (68843 rows, 30 columns, 39.38 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_03000-03250_HITEMP2010.h5 in 1.02s (125728 rows, 30 columns, 71.92 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_03250-03500_HITEMP2010.h5 in 3.75s (465111 rows, 30 columns, 266.04 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_03500-03750_HITEMP2010.h5 in 4.55s (570543 rows, 30 columns, 326.35 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_03750-04000_HITEMP2010.h5 in 1.34s (167996 rows, 30 columns, 96.09 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_04000-04500_HITEMP2010.h5 in 0.79s (99542 rows, 30 columns, 56.94 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_04500-05000_HITEMP2010.h5 in 4.52s (543064 rows, 30 columns, 310.63 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_05000-05500_HITEMP2010.h5 in 3.05s (372220 rows, 30 columns, 212.91 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_05500-06000_HITEMP2010.h5 in 1.02s (127110 rows, 30 columns, 72.71 MB)
Generated dataframe from C:\Users\erwan\.radisdb\CO2-02_06000-06500_HITEMP2010.h5 in 0.08s (3040 rows, 30 columns, 1.74 MB)
...
🏃‍ Database initialized in 34.9s
...
🏃‍ Spectrum calculated in 45.8s (all included)


🏃‍ 2) With Vaex:
Generated dataframe from ['C:\\Users\\erwan\\.radisdb\\CO2-02_01500-02000_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_02000-02125_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_02125-02250_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_02250-02500_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_02500-03000_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_03000-03250_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_03250-03500_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_03500-03750_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_03750-04000_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_04000-04500_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_04500-05000_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_05000-05500_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_05500-06000_HITEMP2010.hdf5', 'C:\\Users\\erwan\\.radisdb\\CO2-02_06000-06500_HITEMP2010.hdf5'] in 12.44s (3512472 rows, 30 columns, 1981.03 MB)
...
🏃‍ Database initialized in 16.9s
...
🏃‍ Spectrum calculated in 28.0s (all included)

```

Of course spectra are the same : 

![image](https://user-images.githubusercontent.com/16088743/128955491-58d8e7de-f083-4b47-9f6f-690d4bd9b33a.png)



So already 2x faster; was 3x faster on my previous test case ; and it will become even much faster when we drop the unnecessary columns at loading.